Fixed #24423 -- Reorganized i18n tag tests.
